added loading products bought and in cart from firebase

+++ app/src/main/java/com/example/shoppingapp/model/User_product_Model.kt
@@ -2,6 +2,6 @@ package com.example.shoppingapp.model
 
 class User_product_Model {
     var key:String?=null
-    var product_id:Float?=null
-    var product_amount:Float?=null
+    var product_id:Int?=null
+    var product_amount:Int?=null
 }
